Heavily refactored namespaces
